# swaggerhub-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ConectaPetAPI",
    "description": "API para o conectaPet",
    "contact": {
      "name": "Jonatan Goulart Maciel",
      "email": "jonatan784@gmail.com"
    },
    "version": "0.3"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor API de Desenvolvimento"
    },
    {
      "url": "https://virtserver.swaggerhub.com/Cryover/ConectaPet/0.3",
      "description": "Servidor API de producao"
    }
  ],
...
  "paths": {
...
    "/usuarios/{id}": {
      "summary": "Represents a user",
      "description": "This resource represents an individual user in the system. Each user is identified by a numeric `id`.\n",
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Busca usuario por ID",
        "description": "A detailed description of the operation. Use markdown for rich text representation, such as **bold**, *italic*, and [links](https://swagger.io).\n",
        "operationId": "getUsuarioById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Atualiza todos campos de usuario.",
        "description": "Atualiza todos campos de usuario.\n",
        "operationId": "updateUsuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Deleta usuario.",
        "description": "Deleta usuario.\n",
        "operationId": "deleteUsuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Atualiza apenas alguns campos.",
        "description": "Atualiza apenas alguns campos ao inves de todos.\n",
        "operationId": "patchUsuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      }
    },
...
    "/consulta/{id}": {
      "summary": "Representa um pet",
      "description": "This resource represents an individual user in the system. Each user is identified by a numeric `id`.\n",
      "get": {
        "tags": [
          "Consulta"
        ],
        "summary": "Busca usuario por ID",
        "description": "A detailed description of the operation. Use markdown for rich text representation, such as **bold**, *italic*, and [links](https://swagger.io).\n",
        "operationId": "getConsultaByIdPet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Consulta ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Consulta"
        ],
        "summary": "Atualiza todos campos de usuario.",
        "description": "Atualiza todos campos de usuario.\n",
        "operationId": "updateConsulta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Consulta"
        ],
        "summary": "Deleta Consulta.",
        "description": "Deleta consulta.\n",
        "operationId": "deleteConsulta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Consulta"
        ],
        "summary": "Atualiza apenas alguns campos.",
        "description": "Atualiza apenas alguns campos ao inves de todos.\n",
        "operationId": "patchConsulta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pet ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              }
            }
          }
        }
      }
    },
...
    "/despesa/{id}": {
      "summary": "Representa um pet",
      "description": "This resource represents an individual user in the system. Each user is identified by a numeric `id`.\n",
      "get": {
        "tags": [
          "Despesa"
        ],
        "summary": "Busca usuario por ID",
        "description": "A detailed description of the operation. Use markdown for rich text representation, such as **bold**, *italic*, and [links](https://swagger.io).\n",
        "operationId": "getDespesasByIdPet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Historico Despesas ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Despesa"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Despesa"
        ],
        "summary": "Atualiza todos campos de Historico de despesa.",
        "description": "Atualiza todos campos de usuario.\n",
        "operationId": "updateDespesa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Despesa"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Despesa"
        ],
        "summary": "Deleta Despesa.",
        "description": "Deleta Despsa.\n",
        "operationId": "deleteDespesa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Despesa"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Despesa"
        ],
        "summary": "Atualiza apenas alguns campos de Historico de despesa.",
        "description": "Atualiza apenas alguns campos ao inves de todos.\n",
        "operationId": "patchDespesa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pet ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Despesa"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        },
        "description": "Objeto Usuario"
      },
...
      "Consulta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "id_dono": {
            "type": "string"
          },
          "id_pet": {
            "type": "string"
          },
          "criado_em": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "tipo_pet": {
            "type": "string"
          },
          "raca": {
            "type": "string"
          },
          "sexo": {
            "type": "string"
          }
        },
        "description": "Objeto Pet"
      },
...
      "Despesa": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "id_pet": {
            "type": "string"
          },
          "criado_em": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "tipo_pet": {
            "type": "string"
          },
          "raca": {
            "type": "string"
          },
          "sexo": {
            "type": "string"
          }
        },
        "description": "Despesa"
      },
...
    },
...
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      },
      "OpenID": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://example.com/.well-known/openid-configuration"
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
              "read": "Grants read access",
              "write": "Grants write access",
              "admin": "Grants access to admin operations"
            }
          }
        }
      }
    }
  }
}